add more detailes

+++ app/src/main/java/com/evya/myweatherapp/model/dailyweathermodel/Wind.kt
@@ -1,13 +1,13 @@
 package com.evya.myweatherapp.model.dailyweathermodel
 
 
-import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
+import com.google.gson.annotations.Expose
 
 data class Wind(
     @SerializedName("deg")
     @Expose
-    val deg: Double,
+    val deg: Int,
 
     @SerializedName("gust")
     @Expose
